attempt fixup auto builds on dotfiles submodule changes
Even though we had dotfiles/** added in the path changes config,
i think this wasn't triggering builds because in the commit, it doesn't
seem to actually record any of the children of dotfiles as changing,
just the dir itself.

Also adding the docker-publish yaml so that changes to it should kick
off a build and publish as well.